Refactor API into package with modular routes
