+++ rockauto/routes/parts.py
@@ -1,342 +1,12 @@
-from fastapi import FastAPI, Query
-
-from pydantic import BaseModel
-from typing import Optional, List, Dict
-
-import random
-
+from fastapi import Query
+from typing import Optional
 import mechanize
 from bs4 import BeautifulSoup
 import html5lib
-
 import requests
-import json
 
-tags_metadata = [
-    {"name": "General", "description": "Welcome message and project overview"},
-    {"name": "Catalog", "description": "Navigate makes, years, models and categories"},
-    {"name": "Parts", "description": "Retrieve parts, closeouts and search the catalog"},
-    {"name": "Vehicles", "description": "Vehicle specific information"},
-]
-
-rockauto_api = FastAPI(
-    title="RockAuto API",
-    description="An unofficial API for RockAuto catalog data",
-    version="0.1.0",
-    openapi_tags=tags_metadata,
-)
-
-@rockauto_api.get("/", tags=["General"])
-async def root():
-    return {
-        "message": "Welcome to the RockAuto API",
-        "description": "Unofficial API for accessing RockAuto catalog data",
-        "endpoints": {
-            "makes": "List all vehicle makes",
-            "years/{search_vehicle}": "Get years for a specific make",
-            "models/{search_vehicle}": "Get models for a specific make and year",
-            "engines/{search_vehicle}": "Get engines for a specific make, year, and model",
-            "categories/{search_vehicle}": "Get categories for a specific vehicle",
-            "sub_categories/{search_vehicle}": "Get sub-categories for a specific category",
-            "parts/{search_vehicle}": "Get parts for a specific vehicle and subcategory",
-            "search": "Search for vehicles and parts with flexible filtering options",
-            "closeouts/{carcode}": "Get closeout deals for a specific vehicle",
-            "vehicle_info/{search_vehicle}": "Get detailed information about a specific vehicle",
-        },
-        "note": "Provide the `link` value from the previous endpoint as `search_link` or omit it to let the API generate one automatically",
-    }
-
-
-
-def craft_search_link(
-    search_make: str = None,
-    search_year: str = None,
-    search_model: str = None,
-    search_engine: str = None,
-    search_category: str = None,
-    search_subcategory: str = None,
-    search_part_type: str = None,
-):
-    """Generate a RockAuto catalog link for the provided parameters."""
-
-    browser = mechanize.Browser()
-    link = "https://www.rockauto.com/en/catalog/"
-    try:
-        if not search_make:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("div", attrs={"class", "ranavnode"})
-        for x in soup:
-            if "US" in next(x.children)["value"]:
-                a = x.find("a", attrs={"class", "navlabellink"})
-                if a and a.get_text().lower() == search_make.lower():
-                    link = "https://www.rockauto.com" + str(a.get("href"))
-                    break
-        else:
-            return None
-
-        if not search_year:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("div", attrs={"class", "ranavnode"})[1:]
-        for x in soup:
-            if "US" in next(x.children)["value"]:
-                a = x.find("a", attrs={"class", "navlabellink"})
-                if a and a.get_text() == search_year:
-                    link = "https://www.rockauto.com" + str(a.get("href"))
-                    break
-        else:
-            return None
-
-        if not search_model:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("div", attrs={"class", "ranavnode"})[2:]
-        for x in soup:
-            if "US" in next(x.children)["value"]:
-                a = x.find("a", attrs={"class", "navlabellink"})
-                if a and a.get_text().lower() == search_model.lower():
-                    link = "https://www.rockauto.com" + str(a.get("href"))
-                    break
-        else:
-            return None
-
-        if not search_engine:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("div", attrs={"class", "ranavnode"})[3:]
-        for x in soup:
-            if "US" in next(x.children)["value"]:
-                a = x.find("a", attrs={"class", "navlabellink"})
-                if a and a.get_text().lower() == search_engine.lower():
-                    link = "https://www.rockauto.com" + str(a.get("href"))
-                    break
-        else:
-            return None
-
-        if not search_category:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("a", attrs={"class", "navlabellink"})[4:]
-        for a in soup:
-            if a.get_text().lower() == search_category.lower():
-                link = "https://www.rockauto.com" + str(a.get("href"))
-                break
-        else:
-            return None
-
-        if not search_subcategory:
-            return link
-
-        page = browser.open(link).read()
-        soup = BeautifulSoup(page, features="html5lib").find_all("a", attrs={"class", "navlabellink"})[5:]
-        for a in soup:
-            if a.get_text().lower() == search_subcategory.lower():
-                link = "https://www.rockauto.com" + str(a.get("href"))
-                break
-        else:
-            return None
-
-        if search_part_type:
-            base_url = "https://www.rockauto.com/en/catalog/"
-            url_parts = [
-                search_make.lower().replace(" ", "+"),
-                search_year,
-                search_model.lower().replace(" ", "+"),
-                search_engine.lower().replace(" ", "+"),
-                "",
-                search_category.lower().replace(" ", "+"),
-                search_subcategory.lower().replace(" ", "+"),
-                search_part_type,
-            ]
-            link_parts = link.split("/")
-            if len(link_parts) >= 6:
-                for part in link_parts:
-                    if part.isdigit() and len(part) > 5:
-                        url_parts[4] = part
-                        break
-            link = base_url + ",".join(url_parts)
-
-        return link
-    except Exception:
-        return None
-    finally:
-        browser.close()
-
-@rockauto_api.get("/makes", tags=["Catalog"])
-async def get_makes():
-    makes_list = []
-
-    browser = mechanize.Browser()
-    page_content = browser.open('https://www.rockauto.com/en/catalog/').read()
-
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('div', attrs={'class', 'ranavnode'})
-    soup_filter = []
-
-    # Find US Market Only
-    for x in soup:
-        if 'US' in next(x.children)['value']:
-            soup_filter.append( x.find('a', attrs={'class', 'navlabellink'}) )
-
-    # Get [Make, Year, Model, Link]
-    for x in soup_filter:
-        makes_list.append( {'make': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return makes_list
-
-@rockauto_api.get("/years/{search_vehicle}", tags=["Catalog"])
-async def get_years(
-    search_make: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/makes` endpoint"),
-):
-    years_list = []
-
-    if not search_link:
-        search_link = craft_search_link(search_make=search_make)
-    browser = mechanize.Browser()
-    page_content = browser.open( search_link ).read()
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('div', attrs={'class', 'ranavnode'})[1:]
-    soup_filter = []
-
-    # Find US Market Only
-    for x in soup:
-        if 'US' in next(x.children)['value']:
-            soup_filter.append( x.find('a', attrs={'class', 'navlabellink'}) )
-
-    # Get [Make, Year, Model, Link]
-    for x in soup_filter:
-        years_list.append( {'make': search_make, 'year': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return years_list
-
-
-@rockauto_api.get("/years/{search_vehicle}", tags=["Catalog"])
-async def get_models(
-    search_make: str,
-    search_year: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/years` endpoint"),
-):
-    models_list = []
-
-    if not search_link:
-        search_link = craft_search_link(search_make=search_make, search_year=search_year)
-    browser = mechanize.Browser()
-    page_content = browser.open( search_link ).read()
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('div', attrs={'class', 'ranavnode'})[2:]
-    soup_filter = []
-
-    # Find US Market Only
-    for x in soup:
-        if 'US' in next(x.children)['value']:
-            soup_filter.append( x.find('a', attrs={'class', 'navlabellink'}) )
-
-    # Get [Make, Year, Model, Link]
-    for x in soup_filter:
-        models_list.append( {'make': search_make, 'year': search_year, 'model': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return models_list
-
-@rockauto_api.get("/engines/{search_vehicle}", tags=["Catalog"])
-async def get_engines(
-    search_make: str,
-    search_year: str,
-    search_model: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/models` endpoint"),
-):
-    engines_list = []
-
-    if not search_link:
-        search_link = craft_search_link(
-            search_make=search_make,
-            search_year=search_year,
-            search_model=search_model,
-        )
-    browser = mechanize.Browser()
-    page_content = browser.open( search_link ).read()
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('div', attrs={'class', 'ranavnode'})[3:]
-    soup_filter = []
-
-    # Find US Market Only
-    for x in soup:
-        if 'US' in next(x.children)['value']:
-            soup_filter.append( x.find('a', attrs={'class', 'navlabellink'}) )
-
-    # Get [Make, Year, Model, Link]
-    for x in soup_filter:
-        engines_list.append( {'make': search_make, 'year': search_year, 'model': search_model, 'engine': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return engines_list
-
-@rockauto_api.get("/categories/{search_vehicle}", tags=["Catalog"])
-async def get_categories(
-    search_make: str,
-    search_year: str,
-    search_model: str,
-    search_engine: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/engines` endpoint"),
-):
-    categories_list = []
-
-    if not search_link:
-        search_link = craft_search_link(
-            search_make=search_make,
-            search_year=search_year,
-            search_model=search_model,
-            search_engine=search_engine,
-        )
-    browser = mechanize.Browser()
-    page_content = browser.open( search_link ).read()
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('a', attrs={'class', 'navlabellink'})[4:]
-
-    for x in soup:
-        categories_list.append( {'make': search_make, 'year': search_year, 'model': search_model, 'engine': search_engine, 'category': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return categories_list
-
-@rockauto_api.get("/sub_categories/{search_vehicle}", tags=["Catalog"])
-async def get_sub_categories(
-    search_make: str,
-    search_year: str,
-    search_model: str,
-    search_engine: str,
-    search_category: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/categories` endpoint"),
-):
-    sub_categories_list = []
-
-    if not search_link:
-        search_link = craft_search_link(
-            search_make=search_make,
-            search_year=search_year,
-            search_model=search_model,
-            search_engine=search_engine,
-            search_category=search_category,
-        )
-    browser = mechanize.Browser()
-    page_content = browser.open( search_link ).read()
-    browser.close()
-
-    soup = BeautifulSoup(page_content, features='html5lib').find_all('a', attrs={'class', 'navlabellink'})[5:]
-
-    for x in soup:
-        sub_categories_list.append( {'make': search_make, 'year': search_year, 'model': search_model, 'engine': search_engine, 'category': search_category, 'sub_category': x.get_text(), 'link': 'https://www.rockauto.com' + str( x.get('href') ) })
-
-    return sub_categories_list
+from .. import rockauto_api
+from .catalog import craft_search_link
 
 @rockauto_api.get(
     "/parts/{search_vehicle}",
@@ -1269,74 +939,3 @@ async def search_part_by_number(partnum: str):
         return results
     except Exception:
         return []
-
-@rockauto_api.get("/vehicle_info/{search_vehicle}", description="Get detailed information about a specific vehicle", tags=["Vehicles"])
-async def get_vehicle_info(
-    search_make: str,
-    search_year: str,
-    search_model: str,
-    search_engine: str,
-    search_link: Optional[str] = Query(None, description="Navigation link from the `/parts` endpoint"),
-):
-    vehicle_info = {
-        'make': search_make,
-        'year': search_year,
-        'model': search_model,
-        'engine': search_engine,
-        'details': {}
-    }
-    
-    if not search_link:
-        search_link = craft_search_link(
-            search_make=search_make,
-            search_year=search_year,
-            search_model=search_model,
-            search_engine=search_engine,
-        )
-    browser = mechanize.Browser()
-    page_content = browser.open(search_link).read()
-    browser.close()
-    
-    soup = BeautifulSoup(page_content, features='html5lib')
-    
-    try:
-        # Get vehicle header info
-        vehicle_header = soup.find('div', attrs={'id': 'vehicleheader'})
-        if vehicle_header:
-            # Extract general vehicle details
-            vehicle_info['details']['header'] = vehicle_header.get_text().strip()
-            
-        # Try to find assembly information
-        assembly_info = soup.find('span', string=lambda text: text and "Assembly" in text)
-        if assembly_info and assembly_info.parent:
-            vehicle_info['details']['assembly'] = assembly_info.parent.get_text().strip()
-        
-        # Try to find body type information
-        body_info = soup.find('span', string=lambda text: text and "Body" in text)
-        if body_info and body_info.parent:
-            vehicle_info['details']['body_type'] = body_info.parent.get_text().strip()
-        
-        # Get engine details
-        engine_info = soup.find('span', string=lambda text: text and "Engine" in text)
-        if engine_info and engine_info.parent:
-            vehicle_info['details']['engine_details'] = engine_info.parent.get_text().strip()
-        
-        # Extract any additional specification table data
-        spec_tables = soup.find_all('table', attrs={'class': 'spec-table'})
-        if spec_tables:
-            specs = {}
-            for table in spec_tables:
-                rows = table.find_all('tr')
-                for row in rows:
-                    columns = row.find_all('td')
-                    if len(columns) >= 2:
-                        key = columns[0].get_text().strip()
-                        value = columns[1].get_text().strip()
-                        specs[key] = value
-            
-            vehicle_info['details']['specifications'] = specs
-    
-    except Exception as e:
-        vehicle_info['error'] = "Error parsing vehicle information"
-    
-    return vehicle_info
